Extract the rewrite generators into a next-package factory
createRewrites() produces the slug/locale rewrites and redirects from
an app's i18n config and slug tables; the only real difference
between apps — SK and MK serving calculators on unprefixed URLs —
becomes an explicit unprefixedCalculatorRoutes flag. Each app's
config/i18n-routing.ts shrinks to a factory call with the same
exported surface for next.config.ts.
