+++ packages/next/src/routing/index.ts
@@ -1,3 +1,4 @@
+export * from "./create-rewrites";
 export * from "./create-routing";
 export { buildCanonicalUrl } from "./factories/url-builders";
 export * from "./guards";
